safari-600.7-branch/LayoutTests/fast/events/platform-wheelevent-paging-y-in-scrolling-page-expected.txt

PASS event.wheelDeltaY is window.givenScrollTop * -3
PASS event.wheelDeltaX is 0
PASS event.wheelDelta is window.givenScrollTop * -3
PASS document.body.scrollTop is window.expectedScrollTop
PASS document.body.scrollLeft is window.expectedScrollLeft
